+++ src/main/java/com/example/deliveryplatform/common/exception/handler/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.example.deliveryplatform.common.exception;
+package com.example.deliveryplatform.common.exception.handler;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -11,6 +11,10 @@ import org.springframework.validation.FieldError;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+import com.example.deliveryplatform.common.exception.BaseException;
+import com.example.deliveryplatform.common.exception.code.ErrorCode;
+import com.example.deliveryplatform.common.exception.dto.ErrorResponse;
 
 import lombok.extern.slf4j.Slf4j;
