+++ src/main/java/com/example/deliveryplatform/domain/auth/service/AuthService.java
@@ -5,7 +5,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.example.deliveryplatform.common.exception.BaseException;
-import com.example.deliveryplatform.common.exception.ErrorCode;
+import com.example.deliveryplatform.common.exception.code.ErrorCode;
 import com.example.deliveryplatform.domain.auth.dto.SignupRequest;
 import com.example.deliveryplatform.domain.auth.dto.SignupResponse;
 import com.example.deliveryplatform.domain.user.entity.User;
